+++ codex-rs/state/src/runtime.rs
@@ -75,7 +75,6 @@ pub use goals::GoalUpdate;
 pub use memories::MemoryStore;
 pub use recovery::RuntimeDbBackup;
 pub use recovery::backup_runtime_db_for_fresh_start;
-pub use recovery::backup_runtime_dbs_for_fresh_start;
 pub use recovery::is_sqlite_corruption_error;
 pub use recovery::runtime_db_path_for_corruption_error;
 pub use recovery::sqlite_error_detail_is_corruption;
